Fix shebangs and x bits on all scripts.
Many scripts had a #! line below the license text, where it has no
effect: the kernel only looks for #! at the very start of the file.
Without that, an executable script will only run successfully if the
shell it's launched from contains fallback processing of some kind for
when the exec system call fails.

Now all bash and Python scripts have an appropriate #! at the very
start, and all of them are consistently marked executable (a few were
missing that as well).