Check Python version before continuing in setup.sh
We currently require 3.5, which happens to be the default for
Ubuntu Xenial:
python3/xenial,now 3.5.1-3 amd64 [installed,automatic]

Also add an intermeddiate var for the python3 command so
it can be something other than "python3".
This isn't so important now, but Python3 is EOL:
DEPRECATION: Python 3.5 reached the end of its life on September 13th,
2020.

So we'll probably move to 3.6 at some point. Meaning
people who install that from an extra repo won't have
it as "python3" but "python3.6". This saves them some hassle.

Ideally you'd be able to tell venv to use that new Python.
However when you use the "python -m venv" style it will only
use the python it's running on.