support building 8m.main and 8.1m.main arches
Three variables were added to config.sh to support multiple configurations. They
are passed to the CFLAGS and ASM_FLAGS options of the diferent Makefiles and are
used to infer certain properties:

- ARCH : What sub-architecture do we want to build. This variable is used both
  in the clang `-target` as in the `-march` cmdline options
- ARCH_OPTIONS : this string is appended to the `-march` architecture option
- FLOAT_ABI : used to populate `-mfloat-abi`

There were several options to choose from to pass these configuration options to
Makefiles. The ARCH_OPTIONS could for our purposes be appended to the target
arch. But this becomes somewhat inflexible. For example compiler-rt dispatches
on architecture name. So we would need to add as many architectures to
compiler-rt as there are options.

I also considered using `-mfpu` to pass floating point options, but it looks
like `-mfpu` can't capture all options properly and the Arm documentation
explicitly warns against using it for armv8.

This setup works for all targets. If we're happy with the compiler-rt changes,
I'll upstream the compiler-rt v8 support to llvm.

I haven't put effort into making some kind of frontend to group and narrow down
compiler options in this patch. I'll do that in a follow-up patch.