In newlib, make sure we don't build math fp assembly with hard fp-abi
Currently, the --enable-newlib-hw-fp flag is queried to see if we should use hw
fp instructions. However, the files in libm/machine/arm will only check compiler
macros to deduce if they should use hw fp or not.

This is not enough when used with MVE because MVE integer mode without general
fp unit will use the fp registers but only to pass integers back and fro.

I think the best way to handle this is to honor --enable-newlib-hw-fp in
libm/machine/arm just as the C level functions do.