Add special case for Clang in deciding floating-point endianness
newlib's ieeefp.h determines floating-point endianness by checking the
__VFP_FP__ macro. If this macro is defined, then newlib assumes that the
target architecture uses the VFP architecture for FPs. In this sort of
arch, IEEE FP values have their word-level endianness the same as their
byte-level endianness.

For example, a double-precision FP value of 1 would be represented in
memory like this (in little-endian ARM targets):

<- lowest addr    highest addr ->
1  0  0  0 | 0  0  0  0

Otherwise, i.e. if __VFP_FP__macro is undefined, then newlib assumes the
FPA architecture for FPs instead. In this case, IEEE FP values are
always big-endian at word level, while at the byte-level they follow the
target architecture's endianness.

For example, the same double-precision FP value of 1 would be
represented like this (in little-endian ARM targets):

<- lowest addr    highest addr ->
0  0  0  0 | 1  0  0  0

The problem is: __VFP_FP__ cannot be used to determine that if using
Clang. In this compiler, the property:

__VFP_FP__ is defined if and only if the VFP architecture is in use.

is false. That's because Clang only defines __VFP_FP__ if an FPU is
enabled, which is not the case when, for instance, soft-float ABI is in
place.

This patch adds another preprocessor check to ieefp.h. Now, if the
compiler used is Clang, newlib can assume the endianness defined by VFP.