redid newlib patch for hardfp libm failures
Previously, as suggested on the newlib website, we used the
`--enable-newlib-hw-fp` flag to toggle between libm directories and also used
this flag to not use Arm floating point instructions.

However it turns out that that alternate fp libm directory is part of an
experiment that has been abandoned. The right way to address this is to check
for the availability of floating point in the C file itself. Also because this
won't be disruptive for current uses of the library.