Add tests to the build process
This patch adds a 'test' phase to the build system.

For now we have added one smoke test, which is the same as
baremetal-semihosting from the 'samples' directory.

If qemu-arm is present in PATH, tests are built and run. Otherwise,
they are only built.

Output checking is done using golden files. Each smoke test must have
one stdout and one stderr file that contain the expected outputs, which
are in turn checked against the test output.

This current test framework requires 'make' installed and in PATH.