Reintroduce sysroot to configuration files
The default C++ include dir path, for bare metal targets, takes into
account only the target triple (which is part of the default sysroot).

Since headers are installed under directories with the same name as
their library variants, it is necessary to specify the correct C++
include path in configuration files.

Given that both the library and C include paths were already fed via the
command-line, and we'd also need to do so for C++ include paths, it's
wiser to reintroduce --sysroot, which makes all these paths work with a
single command-line option instead of three.