Add (experimental) aarch64-none-elf support
This also results in the built clang compiler not having a default
target triple if support is asked for both Arm and AArch64 targets.

This in turn requires following changes:

(1) Always provide a CXX compiler target so that the cmake step that checks
    the C++ compiler can succeed now that a --target option needs to be
    supplied;
(2) Always explicitly specify the --target flag for building the
    libraries.

Only the necessary LLVM backends will be built (ARM and/or AArch64),
based on requested toolchain variants.