Add builds based on LLVM 13
This patch adds two new versions to versions.yml:
- 13.0.0 - based on LLVM 13.0.0 and newlib 4.1.0
- branch-13 - based on the tip of the LLVM 13 branch and newlib 4.1.0

The patch for LLVM 13 is identical to the HEAD patch. The patch for
newlib is based on the exiting newlib-0.1.patch, except for the
newlib/libc/include/machine/ieeefp.h part (it is not needed anymore
because Clang 13 correctly defines the __VFP_FP__ macro).