Add notices to modified LLVM files
The Apache License v2.0 with LLVM Exceptions clause 4.b says:

"You must cause any modified files to carry prominent notices stating
that You changed the files;"

Previously we only distributed the build scripts, and not the modified
LLVM source code. Since we are going to create a source distribution
which includes the patched LLVM source code we need to add the notices
mentioned in the license to our LLVM patches.