Implement building from source package
Currently the build scripts require the LLVM and newlib sources to
be git repositories (e.g. the scripts query git commit hashes). This
patch adds a new mode for building the toolchain from a source package
in which the scripts, LLVM and newlib are bundled together.

The file versions.yml is now used for distinguishing between the two
modes. When the top-level entry 'SourceType' is set to
'standalone-build-scripts' the scripts will assume that LLVM and
newlib need to be checked out from git. Otherwise, if the value is set
to 'source-package' the scripts will assume that LLVM and newlib
sources are located in the same directory as the versions.yml file.

When building a source package the scripts generate versions.yml and
set SourceType to 'source-package'. Also, when build from a source
package the file VERSION.txt is copied from the source package rather
than generated on the fly.

The following command-line options are incompatible with the
'source-package' mode:
- prepare
- package-src
- --repositories-dir
- -r/--revision
If any of these options is specified when build from a source package
the scripts will ignore the option and print a warning.