Add Mingw-w64 licenses and version
This patch adds several license files pertinent to the runtime DLLs
that are part of the Mingw-w64 project, specifically:
* COPYING3.GCC - GPL v3 - applies to libstdc++ and libgcc
* COPYING.RUNTIME - GCC Runtime library exception - applies to
  libstdc++ and libgcc
* COPYING.MinGW-w64-runtime.txt - applies to libstdc++, libgcc and
  libwinpthread

These licenses are included in the binary distribution when the
corresponding DLLs are included in the distribution.

Also this patch adds Mingw-w64 version information to VERSION.txt
when compiling with Mingw-w64.