Use hard-coded commit hashes to build the HEAD version
This is the first patch in a series of patches to fix the build
failure of the HEAD version.

The patch changes commit hashes that are used for building the HEAD
version to hard-coded constants. It also fixes the build scripts to
work correctly when checking out source code based on commit hashes
rather than branches or tags.

The plan is to go through all breaking changes and fix them one by
one updating the commit hashes at the same time, and then revert back
to using the latest commit.