Use new syntax to specify bin directory in config files
Upstream patch https://reviews.llvm.org/D115604 implemented a feature
that we used to implement in a downstream patch. Config files can now
use a special placeholder that is automatically replaced with the
directory containing the config file. The placeholder used by upstream
is different from the downstream one ('<CFGDIR>' vs '@'), this
requires some adjustments in the python script that generates the
config files.