Use newlib commit that does not require autotools
The newlib commit we are currently using has stale
libgloss/aclocal.m4 file, and because of this build fails on machines
that don't have autotools installed.

This patch fixes the issue by changing newlib commit to point to a
commit in which files generated by autotools have been been updated.