Move default install location into build dir
The smoke tests use the install dir, which meant that deleting the build
dir didn't assure a valid test result.
When compared to other build systems this is surprising.
By moving the install dir into the build dir the expectation that
deleting the build dir will cause tests to use fresh build output is
upheld.

For consistency, also default to putting the package tar.gz in the build
directory.