Picolibc (#123)
Replace newlib with picolibc

GitHub issue #61

Change "rdimon" names to use "semihost" instead

Turn off wide character support in libc++ to make it play nicely with
picolibc's tinystdio.

Add a mbstate_t header to picolibc in the place that libc++ expects to
find it when wide character support is disabled.

The ldscript/base_arm.ld linker script doesn't work with picolibc. Maybe
it should be removed entirely but that's a discussion for another time.
For now add the minimal arm.ld in the smoketests directory.

Picolibc's semihosting uses a single FILE object for stdout/err so we
can't discriminate between the two.