Infrastructure to support 32-bit A-class targets.
At the moment, if you try to add a library variant targeting armv7a,
the compiler-rt sub-build will silently not build any actual builtins
library, because when compiler-rt is given a triple like
'armv7a-none-eabi', it doesn't find 'armv7a' in its list of recognised
Arm variants, so it doesn't enable any Arm-related support functions.

A workaround is to put plain 'arm' in the target triple, and specify
the arch version using -march=armv7a. However, _that_ doesn't work for
v6-M, because compiler-rt needs to include a whole different set of
assembly files for that, and if you don't put 'armv6m' in the target
triple, it won't notice that it needs to make that change.

So in this change I arrange to only include the specific arch version
in the target triple if it's absolutely necessary for something to
work, which as far as I know currently means 'only if it's armv6m'.
For everything else, the target triple just says 'arm'.

As far as I know that change is only needed for compiler-rt. But it
doesn't seem to have done any harm to put it in everywhere.

(Another way to fix this would be to ensure every possible
architecture version appears in compiler-rt's list, but that would
involve a lot more ongoing work.)

Also, varying the target triple independently of the architecture
gives the ability to pass 'eabihf' as the ABI component, which is how
you select the hard float ABI in compiler-rt.

There should be no functional change in this commit for any existing
variant that we build. But it removes an obstacle to adding v7-A (or
v8-A, etc) variants in the future.