Allow copying libraries from another build
Building the libraries on Windows can take hours, compared to minutes on
Linux. Therefore this change allows copying libraries from another
build.

The new PREBUILT_TARGET_LIBRARIES CMake option causes the building of
libraries to be skipped, with the assumption that the libraries will
be added in the location in which they would normally be built.

The copy_target_libraries.py script can be used to copy the libraries
from a previously built package.