Generate and install a multilib.yaml file.
The multilib system added to clang by the previous commit doesn't do
anything unless a YAML configuration file is available at run time.
This commit adds one.

This makes the .cfg files optional, although they're still available
to be used if a user prefers. For example, these two commands have the
same effect of building a semihosting hello-world program for Armv7-M,
with and without a config file:

  clang --config=armv7m_soft_nofp_semihost.cfg -T picolibc.ld hello.c
  clang --target=arm-none-eabi -march=armv7-m -T picolibc.ld -lcrt0-semihost -lsemihost hello.c

The multilib.yaml is built by a CMake configure_file command, starting
from a template containing a static FlagMap section, and filling in
the Variants section to match the set of library variants actually
being built. Each library must now provide a set of multilib selection
flags along with the compile flags for the library itself.

Documentation for the YAML format can be found in
clang/docs/Multilib.rst, which is included in the previous commit's
patch update, or alternatively in https://reviews.llvm.org/D143587.