Remove paths to sysroot & crt0 from config files
These are not necessary since the multilib system will find the correct
paths automatically.

This necessitates switching from using crt0.o and crt0-semihost.o to
instead using libcrt0.a and libcrt0-semihost.a. This is because .o
files cannot be found in the linker search directories, whereas .a
files can.