Multilib
This change adds multilib configuration in accordance with the
documentation in https://reviews.llvm.org/D143587

For the purposes of testing the multilib configuration, this change adds
lit test configuration. It is intended that this test configuration will
make it easier to add tests for other purposes as well, instead of
adding targets in directly in CMake.