Run package tests with llvm-lit
Now the output of the tests is checked.
It also makes failing tests easier to debug.

Previously these tests were run on both build directory and the
directory into which the toolchain has been installed. The benefit of
this was questionable, so now those tests are only run on the latter.