Multilib: autogenerate implications between FPU types.
This replaces the previous hand-written set of Match: rules with a
full set that includes every subset/superset relation between any two
FPU types known to clang.

This set of relations is autogenerated at build time, by querying the
clang source for the list of FPU names and querying clang itself for
the list of -target-feature commands implied by each one.

(As a result, multilib.yaml is now a file built by a target, rather
than made by configure_file at cmake time. The output of
configure_file now goes to a temporary file, which is concatenated to
the autogenerated half of the script.)

I've added a new test, which checks all the implications between FPUs
that were previously hard-coded (since those are the ones that the
library variants in this project depend on), and a few selected other
ones. I haven't tested every single implication, because part of the
aim of automating this is to cope automatically with changes to LLVM's
FPU taxonomy, and if our test embodies every single expected result,
then we'll have to fix it anyway every time that changes.