Speed up build time
Each library configure step is typically single-threaded so it's
desirable to do it in parallel with other tasks. Disabling terminal
access for the configure step means that the step will no longer be
placed in a Ninja job pool, and therefore can be run in parallel.

This change speeds up the overall build time by 6 minutes.