add ExclusiveGroup feature to multilib.yaml.
This patch adds an ExclusiveGroup: stdlibs flag to all variants.
Being in the same ExclusiveGroup would result in the last matching
variant out of multiple matching variants to be selected.

When no -mfpu=none is specified, the compiler internally adds all other
possible fpu settings when searching for matching variants. So for the
no-fpu variant to win, it has to be added to multilab.yaml after all
other fpu variants. So reorder the add_library_variant calls accordingly.