Compiler's -S output and `-save-temps=obj` loses MVE feature which leads to assembler error for Arm V8.1m Cpus. (#313) (#315)
At the moment clang is not generating the correct
FPU directive for Arm V8.1m Mainline CPU's when assembled
by -S. So when compiled with -S output and `-save-temps=obj`
loses MVE feature which leads to assembler error.

There is a mismatch in the default FPU option that Arm V8.1-m CPUs
define and what clang generates. For example, Clang generates
fpv5-d16 for M85 and what is defined in ARMTargetParser.def is
FK_FP_ARMV8_FULLFP16_D16.

This patch add cases for handling this FPU directive as well and thus
makes clang generates the right FPU directives for Arm V8.1-m CPUs.