Reintroduce option to use newlib as the C library.
Commit cf09c476134eda7, between versions 14 and 15 of the toolchain,
replaced newlib with picolibc in this toolchain. We now bring back
newlib as an alternative option, although picolibc is still the
default.

User-visible effects:

To build the whole toolchain with newlib-based libraries instead of
picolibc, you can configure with `-DLLVM_TOOLCHAIN_C_LIBRARY=newlib`.
This will cause newlib to be git cloned and compiled in place of
picolibc. The newlib setup is multilib-enabled just as picolibc is, so
you can compile a program with a command line such as

  clang --target=arm-none-eabi -march=armv7m -T redboot.ld -lrdimon -lcrt0-rdimon -o hello hello.c

This differs very little from the corresponding picolibc command line.
The ld script and library names use newlib's naming choices instead of
picolibc's, but the general idea is the same: you say what CPU/FPU/ABI
you want to compile for, and multilib takes care of selecting a
particular build of the libraries.

The `-lrdimon` and `-lcrt0-rdimon` library specifications cause
newlib's semihosting support to be included in the application image.
If you don't want that, you can replace them both with `-lnosys` and
`-lcrt0-nosys`.

If you also set `-DLLVM_TOOLCHAIN_NEWLIB_OVERLAY_INSTALL=on` at cmake
time, then the package-llvm-toolchain target will be modified so that
instead of generating a self-contained toolchain archive, it generates
an overlay archive that can be unpacked over the standard picolibc
archive, to produce a combined toolchain that uses picolibc by default
but can be switched to newlib on demand.

The overlay archive installs all the newlib library variants in a
subdirectory `lib/clang-runtimes/newlib`, so that they don't collide
with the picolibc variants in `lib/clang-runtimes` itself. It also
drops a file `newlib.cfg` into the `bin` directory, which switches the
sysroot to the newlib subdirectory. So if you have a toolchain
containing both libraries in this way, you'd add `--config=newlib.cfg`
to the above command line:

  clang --config=newlib.cfg --target=arm-none-eabi -march=armv7m -T redboot.ld -lrdimon -lcrt0-rdimon -o hello hello.c

Note that the ordinary 'install' target will not work properly if
you've set `-DLLVM_TOOLCHAIN_NEWLIB_OVERLAY_INSTALL=on`. That option
is intended _only_ for generating an overlay archive.

Implementation notes:

This is an initial implementation only. It doesn't arrange to run any
tests of newlib: there's a `check-newlib` build target, but nothing
currently in it.

Picolibc has the option to build its `crt0.o` files into small
libraries, so that the end-user command line doesn't need to state the
full path of the toolchain install directory. Newlib doesn't have that
feature upstream, so instead, the CMakeLists.txt in this project runs
llvm-ar to construct those libraries just after the main `make`
finishes in each newlib directory.

The version of newlib supported is fixed at 4.1.0, matching what was
used by version 14 of this toolchain. Updating it to current newlib is
a separate piece of work not included in this initial patch.

The bulk of `patches/newlib.patch` is dealing with assembly source
files in the newlib source tree which `gas` accepts but clang's
assembler doesn't, because it's more picky about instruction
syntax (e.g. UAL syntax, `p15` rather than plain `15` in coprocessor
instructions, no support for the nonstandard `adrl`
pseudo-instruction).