Read LLVM version from subdirectory
Rather than use a regex string search on the LLVM CMakeLists.txt, this
patch adds the LLVM project as a subdirectory and then reads the
version variables as properties. Additionally, this adds some new
variables to store packaging information, as including CPack resets
many of the variables used to configure CPack.