picolibc: linker script: Add '__lcxx_override' section
This section was added into libc++ and conains the "new" operator[1].
Placing the "new" operator in the section allows checking, if the
operator (defined as weak symbol) was overridden.

The check relies on the "__start_" and "__end_" symbols generated by
linker. These symbols would not be generated for names, which are not
valid C identifiers. For this reason name like ".text.__lcxx_generated"
cannot be used.

[1] https://github.com/llvm/llvm-project/pull/69498