Update target triples affected by normalization change
The way in which clang normalizes certain target triples was
recently changed:

https://github.com/llvm/llvm-project/pull/89638

This has the effect of switching the vendor and OS from
"none-unknown" to "unknown-none" in the normalized triple, to better
reflect that the bare metal target has no OS, rather than an unknown
one.

The cmake script and multilib templates have various hardcoded
references to triples which now need changing.