Change filenames in picolibc test-ungetc-ftell,test-fgets-eof
This adds the patch in https://github.com/picolibc/picolibc/pull/734
to our picolibc patch. We think this is the cause of those two
picolibc tests failing intermittently: they're both trying to make
test output files with the same name in the test directory, and
colliding whenever they happen to run concurrently.