Generalise the 'library overlay install' option.
I've renamed it so it doesn't have NEWLIB in the name: now it's just
called `LLVM_TOOLCHAIN_LIBRARY_OVERLAY_INSTALL`. You configure it to
make an overlay package of newlib in particular by also defining
`LLVM_TOOLCHAIN_C_LIBRARY=newlib` (which you already had to do anyway,
even with the previous option name).

The old option name is no longer a cmake cache entry, but it is
detected on the command line, and used to set the default for the new
option. So existing cmake command lines should still work.

I've also removed some of the literal uses of the string 'newlib' in
code conditionalised by the option. Now those strings are obtained by
expanding `${LLVM_TOOLCHAIN_C_LIBRARY}`. So when we add another libc
type in the next commit, they should automatically adapt.