Add experimental ability to build LLVM libc.
This is extremely new and barely tested: I've compiled and run a
hello-world program, and that is literally all. But that's a start –
if you can do that, you can also investigate what else does and
doesn't compile or run successfully. So this is enough to allow users
to experiment.

As documented in the new `docs/llvmlibc.md`, this currently builds C
libraries only (not C++), for AArch32 only. I don't know how much
effort it will take to fix each of those: I only know that in each
case _at least one_ cause of build failure exists, but I don't know
how many more are hiding behind that one.

You can use the new `-DLLVM_TOOLCHAIN_C_LIBRARY` with or without the
option to build an overlay package. If you have both a newlib and an
llvm-libc overlay package, then it should be possible to install both
alongside each other on top of the same main toolchain.

Unlike picolibc and newlib, llvm-libc doesn't come with a crt0.o
startup file. Compiled for bare metal (as we're doing here), it also
leaves some support functions undefined (e.g. for writing to stdout or
stderr), expecting the user to provide them. So I've also added a
small source directory containing the missing pieces. That builds two
additional libraries, -lcrt0 and -lsemihost.

At present, I haven't set up a llvm-libc specific ld script, and
there's no code in my new crt0 that copies RW data out of ROM (as
there is in picolibc). The only definition required by the crt0 is
`__stack`, which you can define via an ld script of your own if you
want, or directly on the ld.lld command ilne if you prefer.