Fix AArch64 version of __llvm_libc_exit.
I made two mistakes in this function: it used size_t without including
any header file that defines it, and it was indecisive about whether
its argument was called `code` or `status`. But both mistakes only
affected the AArch64 branch of the #ifdef, which we weren't using yet,
so they didn't cause a compile failure.