Build llvmlibc variants for AArch64.
We couldn't previously build AArch64 llvmlibc, because the build
failed due to lack of setjmp/longjmp. But now llvmlibc does have those
functions for AArch64, so we can make building AArch64 library
variants unconditional again, not dependent on which libc you want.

The other missing thing for AArch64 is a configuration file in
libc/config/baremetal (it only has configurations for two
architectures, and AArch64 isn't one of them). I think the easiest
approach is to reuse the AArch32 configuration, so that we build the
same set of library functions into all our bare-metal libraries, 32-
or 64-bit.