Do not default to linking crt0 when using crt0-semihosting
This commit adds the -nostartfiles option to tests using semihosting.
This is in response to a recent change in clang which modified the
driver to include crt0.o in the link:
https://github.com/llvm/llvm-project/pull/101258

While the change to clang removes the need to include the crt0
library directly, the tests using semihosting need to link against a
different crt0 library. Therefore the new default behavior needs to
be suppressed in these cases, which can be done by adding
the -nostartfiles option.