llvmlibc: make time_t 64-bit, even on AArch32.
llvmlibc used to use a 32-bit time_t unconditionally on AArch32,
commented as being for compatibility with glibc. But now it has an
option to choose 64-bit instead, if you don't need to conform to
glibc's ABI.

A bare-metal libc indeed has no need to worry about glibc's time_t
ABI (even for so long as it _stays_ 32-bit, which is probably not much
longer). So it makes sense for us to use 64-bit time_t from the
start, by setting the new config option.