Update comment about the llvmlibc tests.
llvmlibc's own test suite is written in C++, just like the library
itself. But unlike the library, the test suite also depends on having
a C++ library available: it uses iostreams, fstreams, vectors and so
on. So until it's possible to build libc++ on top of a bare-metal
version of llvmlibc, I can't see how we'd be able to run the tests.