Merge v7-M and v7E-M library variants
The only differece between v7-M and v7E-M is the DSP extension, which is
not useful for the C/C++ libraries, so we can simplify things by just
having v7-M builds. This fills in a gap of (v7-M, fpv4-sp-d16,
hard-float), which we did not previously have a valid library for.