Revert "Add Armv8.1-M PACBTI library variants (#467)"
The work done in LLVM to add the `-mbranch-protection` command line
option to the set of flags considered by the multilib implementation
when selecting variants wasn't cherry-picked into the release/19.x
branch. This reverts the creation of PACBTI-M variants in the toolchain
for the corresponding branch, as they can't be selected for multilib.

This reverts commits e63111092f4317a4675150d054091a3507c68af0 and
20834fd4f9f7135101301184decb6627c5246157.