Remove xfail on ctime.timespec.compile.pass.cpp
This test was xfailed as older versions of picolibc did not define TIME_UTC.
A definition was eventually added to picolibc and so the xfail was removed.
This happened after LLVM-19 branching, and so the xfail now needs to be
removed to prevent an unexpected pass with more recent picolibc versions.