Cherry pick setjmp/longjump commit and fix LLVM libc namespace
Building the LLVM libc overlay on AArch64 requires a commit adding
setjmp/longjump for AArch64, which did not make it into the LLVM 19 branch:

cf044b4ec3224d65603bb1805d870a3c111bc297
[libc][AArch64] Add an AArch64 setjmp/longjmp. (#101177)

This commit is therefore included as a downstream patch so that the overlay
can continue to be built.

Additionally, the default namespace used for LLVM's libc includes
LLVM_VERSION_SUFFIX, which takes the form "-rcN" on release candidates. This
namespace is used in macro names, where the "-" character is invalid, leading
to build errors. As a workaround, a name for the namespace without the suffix
is set during configuration, so that the default is not used.